--- a/everestpy/src/everest/misc.hpp
+++ b/everestpy/src/everest/misc.hpp
@@ -14,10 +14,14 @@ const std::string get_variable_from_env(const std::string& variable, const std::
 
 class RuntimeSession {
 public:
+    /// \brief Allows python modules to directly pass \p mqtt_settings as well as a \p logging_config
     RuntimeSession(const Everest::MQTTSettings& mqtt_settings, const std::string& logging_config);
 
+    [[deprecated("Consider switching to the newer RuntimeSession() or RuntimeSession(mqtt_settings, logging_config) "
+                 "ctors that receive module configuration via MQTT")]]
     RuntimeSession(const std::string& prefix, const std::string& config_file);
 
+    /// \brief Get settings and configuration via MQTT based on certain environment variables
     RuntimeSession();
 
     const Everest::MQTTSettings& get_mqtt_settings() const {