--- a/everestpy/src/everest/module.hpp
+++ b/everestpy/src/everest/module.hpp
@@ -3,6 +3,7 @@
 #ifndef EVERESTPY_MODULE_HPP
 #define EVERESTPY_MODULE_HPP
 
+#include <chrono>
 #include <deque>
 #include <functional>
 #include <map>
@@ -27,6 +28,11 @@ public:
         if (on_ready_handler) {
             handle->register_on_ready_handler(std::move(on_ready_handler));
         }
+
+        const auto end_time = std::chrono::system_clock::now();
+        EVLOG_info << "Module " << fmt::format(Everest::TERMINAL_STYLE_BLUE, "{}", this->module_id) << " initialized ["
+                   << std::chrono::duration_cast<std::chrono::milliseconds>(end_time - this->start_time).count()
+                   << "ms]";
 
         handle->signal_ready();
     }
@@ -78,6 +84,7 @@ public:
 private:
     const std::string module_id;
     const RuntimeSession& session;
+    const std::chrono::time_point<std::chrono::system_clock> start_time;
     std::unique_ptr<Everest::RuntimeSettings> rs;
     std::shared_ptr<Everest::MQTTAbstraction> mqtt_abstraction;
     std::unique_ptr<Everest::Config> config_;