--- a/include/framework/runtime.hpp
+++ b/include/framework/runtime.hpp
@@ -194,7 +194,7 @@ public:
     explicit ModuleLoader(int argc, char* argv[], ModuleCallbacks callbacks) :
         ModuleLoader(argc, argv, callbacks, {"undefined project", "undefined version", "undefined git version"}){};
     explicit ModuleLoader(int argc, char* argv[], ModuleCallbacks callbacks,
-                          const VersionInformation version_information);
+                          const VersionInformation& version_information);
 
     int initialize();
 };