* contains performance enhancements in:
  * schema loading, config parsing and validation
  * async MQTT config distribution
  * publishing of manifests and module names
  * handling of types of config items where until now the whole schema was published but only the type was needed

* Additionally addresses some clang-tidy issues

* Adds a new ctor for everestpy to allow MQTT based config distribution to be used in everest-testing more easily

* Remove retained topics after startup and add flag that prevents this from happening for introspection


Some quick-and-dirty performance comparisons starting up the config-sil with a recent everest-core main. This mostly shows that config loading got faster (by cleaning up the usage of schema validators) and a decrease in individual module startup time. Interestingly the total startup time is pretty similar which warrants further investigation:

(v0.19.2 vs v0.19.1)

```
manager          :: everest-framework 0.19.2 feature/performance-improvements@v0.19.1-35-g64e9ae2
manager          :: Config loading completed in 188ms
manager          :: Starting 16 modules
api:API          :: Module api initialized [53ms]
auth:Auth        :: Module auth initialized [55ms]
connector_1:Evs  :: Module connector_1 initialized [81ms]
connector_1_pow  :: Module connector_1_powerpath initialized [57ms]
energy_manager:  :: Module energy_manager initialized [44ms]
error_history:E  :: Module error_history initialized [61ms]
ev_manager:EvMa  :: Module ev_manager initialized [47ms]
evse_security:E  :: Module evse_security initialized [64ms]
grid_connection  :: Module grid_connection_point initialized [68ms]
iso15118_car     :: Module iso15118_car initialized [45ms]
iso15118_charge  :: Module iso15118_charger initialized [61ms]
persistent_stor  :: Module persistent_store initialized [45ms]
setup:Setup      :: Module setup initialized [38ms]
slac:SlacSimula  :: Module slac initialized [42ms]
token_provider:  :: Module token_provider initialized [38ms]
token_validator  :: Module token_validator initialized [35ms]
manager          :: Clearing retained topics published by manager during startup
manager          :: 🚙🚙🚙 All modules are initialized. EVerest up and running [1700ms] 🚙🚙🚙
```
vs.
```
manager          :: everest-framework 0.19.1 @v0.19.1
manager          :: Config loading completed in 336ms
api:API          :: Module api initialized [84ms]
auth:Auth        :: Module auth initialized [65ms]
connector_1:Evs  :: Module connector_1 initialized [86ms]
energy_manager:  :: Module energy_manager initialized [76ms]
error_history:E  :: Module error_history initialized [94ms]
ev_manager:EvMa  :: Module ev_manager initialized [89ms]
evse_security:E  :: Module evse_security initialized [96ms]
grid_connection  :: Module grid_connection_point initialized [79ms]
iso15118_charge  :: Module iso15118_charger initialized [87ms]
persistent_stor  :: Module persistent_store initialized [63ms]
setup:Setup      :: Module setup initialized [60ms]
slac:SlacSimula  :: Module slac initialized [70ms]
token_provider:  :: Module token_provider initialized [67ms]
token_validator  :: Module token_validator initialized [65ms]
manager          :: 🚙🚙🚙 All modules are initialized. EVerest up and running [1721ms] 🚙🚙🚙
```

Comparing the amount of threads a module uses shows a measurable decrease (refactoring the error handling to use less topics/handlers will have an even big impact on eg. ErrorHistory https://github.com/EVerest/everest-framework/pull/225 )

(v0.19.2 vs v0.19.1)
```
manager            35
controller          2
api:API            22
auth:Auth          22
connector_1:Evs   121
connector_1_pow    15
energy_manager:     6
error_history:E   119
ev_manager:EvMa    16
evse_security:E     6
grid_connection     7
python3             5
iso15118_charge     8
persistent_stor     5
setup:Setup        12
slac:SlacSimula     7
token_provider:    13
token_validator     5
```
vs
```
manager            35
controller          2
api:API            55
auth:Auth          55
connector_1:Evs   156
connector_1_pow    47
energy_manager:    39
error_history:E   151
ev_manager:EvMa    50
evse_security:E    38
grid_connection    40
python3            37
iso15118_charge    41
persistent_stor    37
setup:Setup        45
slac:SlacSimula    39
token_provider:    45
token_validator    37
```